+++ frontend/public/o1js/bootstrap.js
@@ -1,3 +1,0 @@
-import * as index from "./src/index.ts";
-
-export default index;