+++ frontend/public/o1js/src/index.ts
@@ -1,6 +1,0 @@
-import { deployZkApp, updateZkApp } from './zk-app/zk-app';
-
-export {
-  deployZkApp,
-  updateZkApp,
-};